Store/python_services/bakong_fastapi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
python-dotenv==1.1.1
requests==2.32.4
